Extract E2E settings module
