+++ tests/e2e/adapters/test_all_adapters.py
@@ -24,7 +24,7 @@ from band_rest import AsyncRestClient
 from band.agent import Agent
 
 from tests.e2e.adapters.conftest import AdapterFactory
-from tests.e2e.conftest import E2ESettings, requires_e2e
+from tests.e2e.settings import E2ESettings, requires_e2e
 from tests.e2e.helpers import (
     TrackingWebSocketClient,
     run_smoke_test,
