+++ tests/e2e/agentcore/test_three_agent_orchestration.py
@@ -36,7 +36,7 @@ import pytest
 from band_rest import AsyncRestClient, CreateMyChatRoomRequestChat
 from band_rest.types import ParticipantRequest
 
-from tests.e2e.conftest import E2ESettings, requires_e2e
+from tests.e2e.settings import E2ESettings, requires_e2e
 from tests.e2e.helpers import (
     TrackingWebSocketClient,
     listening_for_agent_responses,
