+++ tests/e2e/scenarios/agno/test_thoughts.py
@@ -24,7 +24,7 @@ import logging
 import pytest
 from band_rest import AsyncRestClient
 
-from tests.e2e.conftest import E2ESettings, requires_e2e
+from tests.e2e.settings import E2ESettings, requires_e2e
 from tests.e2e.helpers import (
     TrackingWebSocketClient,
     listening_for_room_activity,
