+++ tests/e2e/scenarios/test_context_persistence.py
@@ -23,7 +23,7 @@ from band_rest import AsyncRestClient
 from band.agent import Agent
 
 from tests.e2e.adapters.conftest import AdapterFactory
-from tests.e2e.conftest import E2ESettings, requires_e2e
+from tests.e2e.settings import E2ESettings, requires_e2e
 from tests.e2e.helpers import (
     TrackingWebSocketClient,
     assert_content_contains,
